Separated jq and no jq tests as two files
